+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - support `bit checkout latest` for checkout to the latest version
 - add `--reset` flag to `bit checkout` command for removing local modifications
 - add `--all` flag to `bit checkout` command for executing the checkout on all components
+- add troubleshooting-isolating link to bit status
 
 ## [[13.0.0] - 18.6.2018](https://github.com/teambit/bit/releases/tag/v13.0.0)